Add BaseOneAppPerTest
