+++ docs/manual/working/scalaGuide/main/tests/code/oneapppertest/ExampleSpec.scala
@@ -5,15 +5,16 @@ package scalaguide.tests.scalatest.oneapppertest
 
 import org.scalatest._
 import org.scalatestplus.play._
+import org.scalatestplus.play.guice.GuiceOneAppPerTest
 import play.api.Play
 import play.api.inject.guice._
 
 // #scalafunctionaltest-oneapppertest
-class ExampleSpec extends PlaySpec with OneAppPerTest {
+class ExampleSpec extends PlaySpec with GuiceOneAppPerTest {
 
   // Override newAppForTest if you need an Application with other than
   // default parameters.
-  implicit override def newAppForTest(td: TestData) = new GuiceApplicationBuilder().configure(Map("ehcacheplugin" -> "disabled")).build()
+  override def newAppForTest(td: TestData) = new GuiceApplicationBuilder().configure(Map("ehcacheplugin" -> "disabled")).build()
 
   "The OneAppPerTest trait" must {
     "provide a new Application for each test" in {
