+++ module/src/test/scala/org/scalatestplus/play/AppSpec.scala
@@ -17,9 +17,10 @@ package org.scalatestplus.play
 
 import play.api.test._
 import org.scalatest._
+import org.scalatestplus.play.guice.GuiceOneAppPerTest
 
 /*
  * Play-Test super-suite that provides a new fake App to each test.
 */
-abstract class AppSpec extends WordSpec with MustMatchers with OptionValues with Inside with OneAppPerTest
+abstract class AppSpec extends WordSpec with MustMatchers with OptionValues with Inside with GuiceOneAppPerTest
